#112 introduced UploadTask and UploadRequest base classes. HttpUploadTask extends UploadTask and HttpUploadRequest extends UploadRequest. Made also some internal library refactorings to keep the code base as clean as possible. It's now possible to also set the buffer size used in all the upload tasks.
